fix: allow Review id to be bigger than Integer.MAX (#236)

+++ src/main/java/com/spotify/github/v3/prs/Review.java
@@ -44,7 +44,7 @@ public interface Review {
 
   /** Id. */
   @Nullable
-  Integer id();
+  Long id();
 
   /** User. */
   @Nullable
